// source/frontend/src/models/api/generated/ApiGen_CodeTypes_AcquisitionStatusTypes.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
export enum ApiGen_CodeTypes_AcquisitionStatusTypes {
  ACTIVE = 'ACTIVE',
  ARCHIV = 'ARCHIV',
  CANCEL = 'CANCEL',
  CLOSED = 'CLOSED',
  COMPLT = 'COMPLT',
  DRAFT = 'DRAFT',
  HOLD = 'HOLD',
}